Added Matlab-mex and cpp files
In this commit, I added the c++ and mex functions that were missing from the initial commit.

Regarding added function "compute_cluster.m", it call a mex function which is added, based on c++ codes "compute_networkcluster....cpp" which are also added. I kept versions in a cluster and my own PCs. I am not sure if this one is the last version because now the cluster version is gone.  So this is the version in my PC. Outcome should not be affected. The main change is the datatype, i.e., uint16 or int32, for some input arrays, which I was playing with to optimize memory and computational efficiency, and test different implementation versions. All the versions must render the same outcome, the difference is in the implementation, for example, some of them based on a hash table with linked list or direct access to an index array.